+++ src/components/pairwise-links/orchestrator.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect } from 'react';
 import Row from './row';
 
 function LinksOrchestrator(props) {
