--- src/components/pairwise/index.js
@@ -1,1 +1,0 @@
-export { default as PairwiseLinks } from './links';
